Change frame data types to uint8_t

+++ testbench.h
@@ -1,9 +1,10 @@
 #include <stdio.h>
+#include <stdint.h>
 #include "sad.h"
 
 void run_testbench (
-  int current_frame[240][320],
-  int reference_frame[240][320]
+  uint8_t current_frame[240][320],
+  uint8_t reference_frame[240][320]
 ) {
   int x, y = 0;
   int r, s = 0;
